// src/shared/api/generated/types/ChatHistoryMessage.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ChatHistoryMessage = {
    type: string;
    idMessage: string;
    timestamp: number;
    typeMessage?: string;
    chatId: string;
    textMessage?: string;
    statusMessage?: string;
    sendByApi?: boolean;
    senderId?: string;
    senderName?: string;
    senderContactName?: string;
    deletedMessageId?: string;
    editedMessageId?: string;
    isEdited?: boolean;
    isDeleted?: boolean;
};